Moved Parser functionality into DatePattern and added string to DatePattern to store the parsed expression
