+++ Tests/CronTests/CronTests.swift
@@ -49,7 +49,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJob2() {
-        let pattern = try! parseExpression("10 13 20 9 3 ? *", hash: 0)
+        let pattern = try! DatePattern("10 13 20 9 3 ? *", hash: 0)
         let currentDate = toDate("2016-03-02 09:08:07")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-03-09 20:13:10"))
@@ -58,7 +58,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJob3() {
-        let pattern = try! parseExpression("0 0 12 ? * SUN *", hash: 0)
+        let pattern = try! DatePattern("0 0 12 ? * SUN *", hash: 0)
         let currentDate = toDate("2016-03-02 09:08:07")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-03-06 12:00:00"))
@@ -67,7 +67,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJob4() {
-        let pattern = try! parseExpression("0 0 12 ? * MON *", hash: 0)
+        let pattern = try! DatePattern("0 0 12 ? * MON *", hash: 0)
         let currentDate = toDate("2016-03-02 09:08:07")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-03-07 12:00:00"))
@@ -76,7 +76,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJob5() {
-        let pattern = try! parseExpression("0 0 12 ? * THU *", hash: 0)
+        let pattern = try! DatePattern("0 0 12 ? * THU *", hash: 0)
         let currentDate = toDate("2016-03-02 09:08:07")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-03-03 12:00:00"))
@@ -85,7 +85,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJob6() {
-        let pattern = try! parseExpression("0 0 12 ? * FRI *", hash: 0)
+        let pattern = try! DatePattern("0 0 12 ? * FRI *", hash: 0)
         let currentDate = toDate("2016-03-02 09:08:07")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-03-04 12:00:00"))
@@ -94,7 +94,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJob7() {
-        let pattern = try! parseExpression("56 34 12 31 * * 2016", hash: 0)
+        let pattern = try! DatePattern("56 34 12 31 * * 2016", hash: 0)
         let currentDate = toDate("2016-01-01 23:59:59")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-01-31 12:34:56"))
@@ -108,7 +108,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJob8() {
-        let pattern = try! parseExpression("34 56 12 18 3 * 2016", hash: 0)
+        let pattern = try! DatePattern("34 56 12 18 3 * 2016", hash: 0)
         let currentDate = toDate("2016-03-02 09:08:07")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-03-18 12:56:34"))
@@ -116,14 +116,14 @@ class CronTests: XCTestCase {
     }
 
     func testCronJob9() {
-        let pattern = try! parseExpression("34 56 12 18 3 * 2016", hash: 0)
+        let pattern = try! DatePattern("34 56 12 18 3 * 2016", hash: 0)
         let currentDate = toDate("2016-03-18 12:57:00")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertNil(generator.next())
     }
 
     func testCronJob10() {
-        let pattern = try! parseExpression("15 0 12 * 3 * 2016", hash: 0)
+        let pattern = try! DatePattern("15 0 12 * 3 * 2016", hash: 0)
         let currentDate = toDate("2016-03-18 12:57:00")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-03-19 12:00:15"))
@@ -143,7 +143,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJob11() {
-        let pattern = try! parseExpression("15 0 * 18 3 * 2016", hash: 0)
+        let pattern = try! DatePattern("15 0 * 18 3 * 2016", hash: 0)
         let currentDate = toDate("2016-03-18 20:57:00")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-03-18 21:00:15"))
@@ -153,7 +153,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJob12() {
-        let pattern = try! parseExpression("0 8 * * *", hash: 0)
+        let pattern = try! DatePattern("0 8 * * *", hash: 0)
         let currentDate = toDate("2016-06-09 18:15:00")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-06-10 08:00:00"))
@@ -162,7 +162,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJobDoW() {
-        let pattern = try! parseExpression("34 56 12 * * WED 2016", hash: 0)
+        let pattern = try! DatePattern("34 56 12 * * WED 2016", hash: 0)
         let currentDate = toDate("2016-02-28 12:57:00")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-03-02 12:56:34"))
@@ -174,7 +174,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJobDoW2() {
-        let pattern = try! parseExpression("4 8 9 * * MON-WED 2016", hash: 0)
+        let pattern = try! DatePattern("4 8 9 * * MON-WED 2016", hash: 0)
         let currentDate = toDate("2016-02-08 12:57:00")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-02-09 09:08:04"))
@@ -191,7 +191,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJobDoW3() {
-        let pattern = try! parseExpression("4 8 9 * * MON,WED 2016", hash: 0)
+        let pattern = try! DatePattern("4 8 9 * * MON,WED 2016", hash: 0)
         let currentDate = toDate("2016-02-08 12:57:00")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-02-10 09:08:04"))
@@ -204,7 +204,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJobDoW4() {
-        let pattern = try! parseExpression("4 8 9 4-15 2 MON-WED 2016", hash: 0)
+        let pattern = try! DatePattern("4 8 9 4-15 2 MON-WED 2016", hash: 0)
         let currentDate = toDate("2016-02-08 12:57:00")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2016-02-09 09:08:04"))
@@ -214,7 +214,7 @@ class CronTests: XCTestCase {
     }
 
     func testCronJobLeaps() {
-        let pattern = try! parseExpression("05 43 21 29 2 * *", hash: 0)
+        let pattern = try! DatePattern("05 43 21 29 2 * *", hash: 0)
         let currentDate = toDate("1999-01-11 12:57:00")
         var generator = DateGenerator(pattern: pattern, hash: 0, date: currentDate)
         XCTAssertEqual(generator.next(), toDate("2000-02-29 21:43:05"))
